moved folders, modified more docstrings
