+++ src/thunderbots/software/radio_communication/mrf/usb/libusb_misc.h
@@ -14,7 +14,7 @@
 
 namespace USB {
     /**
-     * These are for internal use only.
+     * Checks a call into libusb if it returned an error, and throws an exception if so.
      */
     long check_fn(const char *call, long err, unsigned int endpoint);
 }
