+++ src/thunderbots/software/util/property.h
@@ -8,35 +8,35 @@
 #include "util/noncopyable.h"
 
 /**
- * \brief A variable holding a value of some type, with the ability to notify
+ * A variable holding a value of some type, with the ability to notify
  * listeners when the value changes.
  *
- * \tparam T the type of value to hold.
+ * @param <T> the type of value to hold.
  */
 template <typename T>
 class Property final : public NonCopyable
 {
    public:
     /**
-     * \brief Constructs a new Property.
+     * Constructs a new Property.
      *
-     * \param[in] value the value with which to initialize the Property.
+     * @param value the value with which to initialize the Property.
      */
     explicit Property(const T &value) : value(value) {}
 
     /**
-     * \brief Move-constructs a new Property.
+     * Move-constructs a new Property.
      *
-     * \param[in] value the value with which to initialize the Property.
+     * @param value the value with which to initialize the Property.
      */
     explicit Property(T &&value) : value(std::move(value)) {}
 
     /**
-     * \brief Move-constructs a Property.
+     * Move-constructs a Property.
      *
      * Properties are not copyable, but they are movable.
      *
-     * \param[in] moveref the original Property to destroy while initializing
+     * @param moveref the original Property to destroy while initializing
      * this Property.
      */
     Property(Property<T> &&moveref)
@@ -47,10 +47,10 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Returns the signal fired when the value of the Property is about
+     * Returns the signal fired when the value of the Property is about
      * to change.
      *
-     * \return the signal.
+     * @return the signal.
      */
     sigc::signal<void> &signal_changing() const
     {
@@ -58,9 +58,9 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Returns the signal fired when the value of the Property changes.
+     * Returns the signal fired when the value of the Property changes.
      *
-     * \return the signal.
+     * @return the signal.
      */
     sigc::signal<void> &signal_changed() const
     {
@@ -68,11 +68,11 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Move-assigns a Property.
+     * Move-assigns a Property.
      *
      * Properties are not copyable, but they are movable.
      *
-     * \param[in] moveref the original Property to destroy while initializing
+     * @param moveref the original Property to destroy while initializing
      * this Property.
      */
     Property<T> &operator=(Property<T> &&moveref)
@@ -84,9 +84,9 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Assigns a new value to the Property.
+     * Assigns a new value to the Property.
      *
-     * \param[in] val the new value to assign.
+     * @param val the new value to assign.
      */
     Property &operator=(const T &val)
     {
@@ -100,9 +100,9 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Move-assigns a new value to the Property.
+     * Move-assigns a new value to the Property.
      *
-     * \param[in] val the new value to assign.
+     * @param val the new value to assign.
      */
     Property &operator=(T &&val)
     {
@@ -116,9 +116,9 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Returns the value of the Property.
+     * Returns the value of the Property.
      *
-     * \return the value.
+     * @return the value.
      */
     const T &get() const
     {
@@ -126,9 +126,9 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Returns the value of the Property.
+     * Returns the value of the Property.
      *
-     * \return the value.
+     * @return the value.
      */
     operator T() const
     {
@@ -136,9 +136,9 @@ class Property final : public NonCopyable
     }
 
     /**
-     * \brief Returns the value of the Property.
+     * Returns the value of the Property.
      *
-     * \return the value.
+     * @return the value.
      */
     const T &operator->() const
     {
